[Chore #141904245] Refactor index imports to fix Travis
