+++ server/controllers/RoleController.js
@@ -1,4 +1,4 @@
-import db from '../models/index';
+import db from '../models/Index.js';
 
 /**
  * Role Controller
